Consolidated review (#90)
* Properly format the title

* add examples

* unfix example name

* Make a temporary code system

* Proper explanation for warning suppression

* Update index.md

Significantly shortened and aligned the 'outline' to match the organization we now have.

* Update background.md

* Update core-concepts.md

Tightening up the text

* Update core-concepts.md

* Update core-concepts.md

* Update workflow-state-overview.md

Tightened up some language. Only looked briefly through the table.

* Update workflow-patterns.md

Added more space

* Update cancelation-when-in-progress-example-task-at-fulfiller.plantuml

Fixing long line

* Update index.md

* Update README.md

* Update index.md

* add scope to BnR

* update text from discussion

* Update background.md

* Update README.md

* add request cancellation

* fix typo

* typo fixes

* Add example

* Update index.md (#82)

Removed unnecessary "or"

* update cancellation

* fix STU note

* Update index.md

* Update index.md

Updated formating in the Boundaries and Relationships Paragraph to be consistent in the paragraph.

* Update examples.md

Removed unnecessary line: |Placer-assigned Lab order| etc. | Placer-assigned | in Example Scenarios Table

* Update order-grouping.md

Removed the line "Order grouping" at the beginn of the File

* Update index.md (#89)

Added a sentence to what "Aspects not detailed in this IG":

Details about the information requirements to execute a workflow (e.g. which Specimen detail information is needed to execute a Laboratory test, which Patient detail Information is needed for a Referral, which detailed information is required regarding a MedicationRequest). This specification only adds constraints to ensure data exchange about workflows is interoperable. Details about needed information are domain-specific and usually jurisdiction-specific, and should be defined in specific Implementation Guides).

* update the descitpions of the cancel examples

* note to balloters about Task.performer

* remove text implying that cancellation could only be for "in-process"

---------

Co-authored-by: brianswinkels <bswinkel1235@gmail.com>
Co-authored-by: Jose Costa Teixeira <jose.a.teixeira@gmail.com>
Co-authored-by: herzogr7 <108152077+herzogr7@users.noreply.github.com>